% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDynamicEnumNodes.R
\name{addDynamicEnumNodes}
\alias{addDynamicEnumNodes}
\title{Get the top nodes for dynamic enum.}
\usage{
addDynamicEnumNodes(curated_col, dd)
}
\arguments{
\item{curated_col}{A character (1). The curated column name to find
the dynamic enum node for}

\item{dd}{A data frame. Data dictionary including the `curated_col` under
its `col.name` column.}
}
\description{
Get the top nodes for dynamic enum.
}
