% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMap.R
\name{convertMap}
\alias{convertMap}
\title{Convert one-to-many map to one-to-one format}
\usage{
convertMap(map, delimiter)
}
\arguments{
\item{map}{A data frame. The original map featuring one-to-many mapping of
original values to curated values in need of updating.}

\item{delimiter}{A character(1). The delimiter used to separate values within
a single cell of the original map. Likely ";" or "<;>".}
}
\value{
A new data frame map featuring the same unique set of original values
with a one-to-one mapping scheme for curated values.
}
\description{
Convert one-to-many map to one-to-one format
}
\examples{
convertMap(cBioPortal_population_ancestry_map, ";")

}
