% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{.bestRoots}
\alias{.bestRoots}
\title{Selects representative nodes for a cluster when LCAs are not available}
\usage{
.bestRoots(graph, vex, ovex)
}
\arguments{
\item{graph}{An igraph network object.}

\item{vex}{Character vector of term ids that make up a single cluster.}

\item{ovex}{Character vector of original term ids present in the cluster.}
}
\value{
Character vector of representative term ids.
}
\description{
Selects representative nodes for a cluster when LCAs are not available
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoCuration")
net <- read.csv(file.path(dir, "sample_net.csv"))

graph <- igraph::graph_from_data_frame(d = net,
vertices = unique(unlist(net)))

vex <- c("NCIT:C78274", "NCIT:C270", "NCIT:C783", "NCIT:C93038",
"NCIT:C47793", "NCIT:C247", "NCIT:C62002")
ovex <- c("NCIT:C270", "NCIT:C93038")

# .bestRoots(graph = graph, vex = vex, ovex = ovex)

}
\keyword{internal}
