% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{.createNetwork}
\alias{.createNetwork}
\title{Transforms rols tree representation into igraph-compatible dataframe network
representation}
\usage{
.createNetwork(tree_frame)
}
\arguments{
\item{tree_frame}{rols JSON tree in dataframe format as created by
jsonlite::fromJSON.}
}
\value{
Dataframe containing a symbolic edge list of a directed network in
the first two columns. Edges are directed from the first column to the second
column. Additional columns are considered as edge attributes.
}
\description{
Transforms rols tree representation into igraph-compatible dataframe network
representation
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoCuration")
tree_frame <- read.csv(file.path(dir, "sample_treeframe.csv"))

# .createNetwork(tree_frame = tree_frame)

}
\keyword{internal}
