% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olsInterface.R
\name{.displayNodes}
\alias{.displayNodes}
\title{Retrieves ontology terms and database information for given term ids}
\usage{
.displayNodes(onto, node)
}
\arguments{
\item{onto}{A character vector. Name(s) of ontologies that terms are from.}

\item{node}{A character vector of ontology term IDs.}
}
\value{
Dataframe of submitted term IDs, term names, and term ontologies
}
\description{
Retrieves ontology terms and database information for given term ids
}
\examples{
onto <- c("FOODON", "CHEBI", "NCIT", "NCIT", "NCIT", "SNOMED", "SNOMED",
"SNOMED")
node <- c("FOODON:03600010", "CHEBI:166822", "NCIT:C1908", "NCIT:C41132",
"NCIT:C25218", "SNOMED:438451000124100", "SNOMED:372740001",
"SNOMED:48070003")

# .displayNodes(onto = onto, node = node)

}
\keyword{internal}
