% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPlotting.R
\name{.drawNetLegend}
\alias{.drawNetLegend}
\title{Draw the standard node-type legend for the network plots}
\usage{
.drawNetLegend(legend_colors, legend_labels)
}
\description{
Draw the standard node-type legend for the network plots
}
\keyword{internal}
