% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillDataDictionary.R
\name{fillDataDictionary}
\alias{fillDataDictionary}
\title{Populate the data dictionary template with the information for a specific attribute(s)}
\usage{
fillDataDictionary(template_dd, attr_dd)
}
\arguments{
\item{template_dd}{A data frame. Template data dictionary containing eight
columns - "col.name", "col.class", "uniqueness", "requiredness",
"multiplevalues", "description", "allowedvalues", "ontology"}

\item{attr_dd}{A data frame. Populated data dictionary for a specific
attribute(s). It should contains the same columns as `template_dd` (i.e.,
the same number and the same name of the columns).}
}
\value{
A data frame, an input for the `template_dd` argument, updated with
the `attr_dd` contents.
}
\description{
Populate the data dictionary template with the information for a specific attribute(s)
}
