% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCuratedData.R
\name{format_list}
\alias{format_list}
\title{Remove duplicate values and clean empty values from a list (SUB-FUNCTION)}
\usage{
format_list(vals_list, delimiter)
}
\arguments{
\item{vals_list}{A list. The list of values belonging to a single cell
to collapse.}

\item{delimiter}{A character(1). The delimiter used to separate values within
a single cell of the original curated data frame. Likely ";" or "<;>".}
}
\value{
A character value containing all values from the list collapsed on the
delimiter.
}
\description{
Remove duplicate values and clean empty values from a list (SUB-FUNCTION)
}
\examples{
format_list(list("red", "yellow", "red", "blue", "", ""), ";")

}
