% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSchema.R
\name{get_required_fields}
\alias{get_required_fields}
\title{Get all required fields from schema}
\usage{
get_required_fields(schema)
}
\arguments{
\item{schema}{A list object returned by \code{\link{load_metadata_schema}}.}
}
\value{
A character vector containing the names of all required fields. Returns
  an empty character vector if no fields are marked as required.
}
\description{
Extracts the names of all fields marked as required in the schema. Required
fields must be present in any dataset validated against this schema.
}
\examples{
schema_file <- system.file("schema", "cmd_schema.yaml",
                          package = "OmicsMLRepoCuration")
if (file.exists(schema_file)) {
  schema <- load_metadata_schema(schema_file)
  required_fields <- get_required_fields(schema)
  print(required_fields)
}

}
