% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntoDefinition.R
\name{getDescription}
\alias{getDescription}
\title{Get the definition of ontology terms}
\usage{
getDescription(term, qryOntoDB = NULL, rows = 10)
}
\arguments{
\item{term}{Character(1). The ontolgy term you want to check its definition.}

\item{qryOntoDB}{The ontology database of the input \code{term} comes from.}

\item{rows}{Integer(1).The maximum number of terms to look up. Default is 10.}
}
\value{
The description/definition of the \code{term} in \code{qryOntoDB}. 
If \code{qryOntoDB} is not specified, the data frame containing all the 
ontology databases including the requested term will be returned.
}
\description{
Get the definition of ontology terms
}
\examples{
getDescription("Sitagliptin", "NCIT")
getDescription("Sitagliptin")


}
