% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnumNodes.R
\name{getEnumNodes}
\alias{getEnumNodes}
\title{Retrieves ideal representative nodes for a vector of ontology term ids}
\usage{
getEnumNodes(ids, cutoff = 0.25)
}
\arguments{
\item{ids}{Character vector of term ids.}

\item{cutoff}{A numeric between 0 and 1. The maximum proportion of terms
covered by the choosen nodes. The 'universe' is all the input terms, the
shared root of them, and the terms between them. The smaller cutoff, the
more nodes will be returned.}
}
\value{
A dataframe of chosen nodes including information on number of
original terms covered.
}
\description{
Retrieves ideal representative nodes for a vector of ontology term ids
}
\examples{
ids <- c("CHEBI:166822", "NCIT:C47639", "FOODON:03600010", "NCIT:C29249",
"NCIT:C983", "NCIT:C247", "NCIT:C47384", "NCIT:C62002", "NCIT:C250",
"NCIT:C270", "NCIT:C94631", "NCIT:C281", "NCIT:C29711", "NCIT:C270")

# getEnumNodes(ids = ids, cutoff = 0.25)

}
