% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getOntologies}
\alias{getOntologies}
\alias{get_ontologies}
\title{Extract ontology from the ontology term id}
\usage{
getOntologies(terms, delim = ":")

get_ontologies(terms, delim = ":")
}
\arguments{
\item{terms}{A character vector}

\item{delim}{A character. Delimiter between ontology and its id.
Default is `:`.}
}
\description{
Extract ontology from the ontology term id
}
\examples{
terms <- c("HP:0001824", "MONDO:0010200", "NCIT:C122328")
getOntologies(terms = terms)

}
