% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getReplacement}
\alias{getReplacement}
\alias{get_replacement}
\title{Get replacement for obsolete term}
\usage{
getReplacement(term)

get_replacement(term)
}
\arguments{
\item{term}{Character; ontology term id}
}
\value{
Character; id of replacement term or "No replacement"
}
\description{
Get replacement for obsolete term
}
\examples{
getReplacement("EFO:0005842")

}
