% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntoDefinition.R
\name{getTermFromOtherDB}
\alias{getTermFromOtherDB}
\title{Get the definition of ontology terms}
\usage{
getTermFromOtherDB(term, targetOntoDB, rows = 10)
}
\arguments{
\item{term}{Character(1). The ontolgy term you want to check its definition.}

\item{targetOntoDB}{A character vector containing the ontology databases you
want to look up the \code{term}. The other ontology databases you want to look for the 
same/similar term as the input \code{term}.}

\item{rows}{Integer(1).The maximum number of terms to look up. Default is 10.}
}
\description{
Get the definition of ontology terms
}
