% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_obsolete}
\alias{is_obsolete}
\title{Indicate if a term is obsolete}
\usage{
is_obsolete(term)
}
\arguments{
\item{term}{Character; ontology term}
}
\value{
Boolean
}
\description{
Indicate if a term is obsolete
}
\examples{
is_obsolete("EFO:0005842")
}
