% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isObsolete}
\alias{isObsolete}
\alias{is_obsolete}
\title{Indicate if a term is obsolete}
\usage{
isObsolete(term)

is_obsolete(term)
}
\arguments{
\item{term}{Character; ontology term}
}
\value{
Boolean
}
\description{
Indicate if a term is obsolete
}
\examples{
isObsolete("EFO:0005842")

}
