% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCuratedData.R
\name{map_values}
\alias{map_values}
\title{Map original values to curated values and curated ontology term ids (SUB-FUNCTION)}
\usage{
map_values(new_map, col, delimiter)
}
\arguments{
\item{new_map}{A data frame. The complete updated map containing the original
values, curated values, and curated term ids for mapping.}

\item{col}{A character(1). The curated column name from the map to pull values 
from (either curated ontology terms or curated ontology term ids).}

\item{delimiter}{A character(1). The delimiter used to separate values within
a single cell of the original curated data frame. Likely ";" or "<;>".}
}
\value{
A character value containing all values fom the list collapsed on the
delimiter.
}
\description{
Map original values to curated values and curated ontology term ids (SUB-FUNCTION)
}
\examples{
map_values(cBioPortal_bodysite_map, "curated_ontology", "<;>")
map_values(cMD_feeding_map, "curated_ontology_term_id", ";")
}
