% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodes.R
\name{mapNodes}
\alias{mapNodes}
\title{Retrieves ideal representative nodes for a vector of ontology term ids}
\usage{
mapNodes(ids, cutoff = 0.25)
}
\arguments{
\item{ids}{Character vector of term ids.}
}
\value{
A dataframe of chosen nodes including information on number of 
original terms covered.
}
\description{
Retrieves ideal representative nodes for a vector of ontology term ids
}
\examples{
ids <- c("CHEBI:166822", "NCIT:C47639", "FOODON:03600010", "NCIT:C29249",
"NCIT:C983", "NCIT:C247", "NCIT:C47384", "NCIT:C62002", "NCIT:C250",
"NCIT:C270", "NCIT:C94631", "NCIT:C281", "NCIT:C29711", "NCIT:C270")

# mapNodes(ids = ids, cutoff = 0.25)

}
