% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCuratedData.R
\name{mapValues}
\alias{mapValues}
\alias{map_values}
\title{Map original values to curated values and curated ontology term ids (SUB-FUNCTION)}
\usage{
mapValues(new_map, col, delimiter, y)

map_values(new_map, col, delimiter, y)
}
\arguments{
\item{new_map}{A data frame. The complete updated map containing the original
values, curated values, and curated term ids for mapping.}

\item{col}{A character(1). The curated column name from the map to pull values 
from (either curated ontology terms or curated ontology term ids).}

\item{delimiter}{A character(1). The delimiter used to separate values within
a single cell of the original curated data frame. Likely ";" or "<;>".}

\item{y}{A character(1). The single original term value which needs to be 
mapped to its curated ontology value.}
}
\value{
A character value containing all values from the list collapsed on the
delimiter.
}
\description{
Map original values to curated values and curated ontology term ids (SUB-FUNCTION)
}
\examples{
ex_df <- data.frame(original_value = c("gold", "gold"), curated_ontology = c("Yellow", "Gold"), curated_ontology_term_id = c("123", "456"))
mapValues(ex_df, "curated_ontology", "<;>", "gold")
}
