% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPlotting.R
\name{markNet}
\alias{markNet}
\title{Plot a network graph with specific vertices highlighted by vertex color and a
a smoothed polygon}
\usage{
markNet(
  net,
  mark_nodes,
  mark_color = "white",
  mark_label = "Marked node",
  layout = layout_with_fr
)
}
\arguments{
\item{net}{An igraph network object with a "type" vertex attribute;
"type" = "root", "bridge", "original", or "picked"}

\item{mark_nodes}{Character vector of vertices to highlight}

\item{mark_color}{Character; color to highlight vertices}

\item{mark_label}{Character; legend label for highlighted vertices; defaults
to "Marked node"}

\item{layout}{An igraph layout object or layout function}
}
\value{
An igraph plot
}
\description{
Plot a network graph with specific vertices highlighted by vertex color and a
a smoothed polygon
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoCuration")
net <- read.csv(file.path(dir, "sample_net.csv"))

netgraph <- igraph::graph_from_data_frame(d = net,
vertices = unique(unlist(net)))
original_terms <- c("NCIT:C29711", "NCIT:C270", "NCIT:C250", "NCIT:C47639",
"NCIT:C29249", "NCIT:C983", "NCIT:C247", "NCIT:C47384", "NCIT:C62002",
"NCIT:C281")

V(netgraph)$type <- "bridge"
V(netgraph)[original_terms]$type <- "original"
V(netgraph)["NCIT:C1908"]$type <- "root"
V(netgraph)["NCIT:C78272"]$type <- "picked"

marked <- c("NCIT:C1909", "NCIT:C1908", "NCIT:C78276", "NCIT:C471",
"NCIT:C78274", "NCIT:C93038")

# markNet(net = netgraph, mark_nodes = marked)

}
