% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxoMap.R
\name{oxoMap}
\alias{oxoMap}
\title{Mapping terms to the same ontology}
\usage{
oxoMap(term_ids, target_ontology = "none", mapping_distance = 1)
}
\arguments{
\item{term_ids}{A character vector of term ids to retrieve mappings for}

\item{target_ontology}{Optional character (1). The target ontology library to
map the `term_ids` to.}

\item{mapping_distance}{Optional. Mapping distance as an integer, 1
(default), 2, or 3.}
}
\value{
A list of of data frames with mapping results for each input term
}
\description{
This function maps input ontology terms to a identical/similar terms in
the target ontology through Ontology Xref Service (OxO) API.
(https://www.ebi.ac.uk/spot/oxo/)
}
\examples{
term_ids <- c("NCIT:C2855", "EFO:0000228", "NCIT:C35025", "HP:0003003",
              "HP:0001824", "MONDO:0010200", "NCIT:C122328")
oxoMap(term_ids, "NCIT")

}
