% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strVsplit}
\alias{strVsplit}
\title{Extract ontology from the ontology term id}
\usage{
strVsplit(terms, delim)
}
\arguments{
\item{terms}{A character vector}

\item{delim}{A character. Delimiter}
}
\value{
A character vector with unique values from the input `terms`
}
\description{
Extract ontology from the ontology term id
}
