% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCuratedData.R
\name{updateCuratedData}
\alias{updateCuratedData}
\title{Update a column of curated data using a new mapping schema}
\usage{
updateCuratedData(curated_data, map, column, delimiter)
}
\arguments{
\item{curated_data}{A data frame. The original version of the curated data
which needs to be updated using the new map.}

\item{map}{A data frame. The new or updated version of the map which 
corresponds to the curation of the selected dataset. The map must be 
formatted in a one-to-one schema of original to curated values.}

\item{column}{A character(1). The root name of the column which needs to be 
updated (excluding "original_" or "curated_").}

\item{delimiter}{A character(1). The delimiter used to separate values within
a single cell of the original curated data frame. Likely ";" or "<;>".}
}
\value{
A data frame of the updated curated data.
}
\description{
Update a column of curated data using a new mapping schema
}
\examples{
ex_map <- data.frame(original_value = c("gold", "gold", "blue", "teal", "teal"), curated_ontology = c("Yellow", "Gold", "Blue", "Teal", "Blue"), curated_ontology_term_id = c("123", "456", "777", "333", "777"))
ex_data <- data.frame(original_color = c("blue", "gold", "teal"), curated_color = c("Blue", "Gold", "Teal"), curated_color_ontology_term_id = c("777", "456", "333"))
updateCuratedData(ex_data, ex_map, "color", "<;>")

}
