% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCurationStats.R
\name{checkCurationStats}
\alias{checkCurationStats}
\title{Calculate the completeness of metadata}
\usage{
checkCurationStats(
  fields_list,
  check = "completeness",
  DB = NULL,
  show_available_fields = FALSE
)
}
\arguments{
\item{fields_list}{A character vector with the original columns/fields names.
Multiple values should be separated by semi-colon(;).}

\item{check}{A character(1) specifying a status of requested fields.
Currently available values are \code{completeness} (default) and
\code{unique}}

\item{DB}{A table you want to check the completeness or uniqueness of
each attributes/columns. The table can be provided in two ways: 1) a
character(1) specifying a database/dataset you want to check the
completeness of fields selected through \code{fields} argument. Or
you can 2) directly provide the metadata table (recommended for a
large metadata such as cBioPortal). Currently, metadata from
curatedMetagenomicData ("\code{cMD}") is only supported.}

\item{show_available_fields}{Default is \code{FALSE}. If is it set to
\code{TRUE}, names of all the available fields/columns will be returned as
a character vector.}
}
\value{
A character vector with the same length as that of \code{fields_list}
representing the percentage completeness of the requested fields.
}
\description{
Calculate the completeness of metadata
}
\examples{
calCompleteness(c("curated_age_years", "curated_age_group"), "cMD")
calCompleteness(c("curated_age_years;curated_age_group"), "cMD")


}
