% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodes.R
\name{.busyNodes}
\alias{.busyNodes}
\title{Detects nodes that cover greater than a certain percentage of the network}
\usage{
.busyNodes(netgraph, max_nodes)
}
\arguments{
\item{netgraph}{igraph network object of ontology term IDs.}

\item{max_nodes}{Integer to indicate number of covered nodes that makes a
node "high-traffic."}
}
\value{
A character vector of high-traffic node ids.
}
\description{
Detects nodes that cover greater than a certain percentage of the network
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoCuration")
net <- read.csv(file.path(dir, "sample_net.csv"))

netgraph <- igraph::graph_from_data_frame(d = net,
vertices = unique(unlist(net)))

# .busyNodes(netgraph = netgraph, max_nodes = 20)

}
\keyword{internal}
