% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodes.R
\name{.clusterNodes}
\alias{.clusterNodes}
\title{Selects representative nodes for a network of ontology terms using igraph
clustering functionality}
\usage{
.clusterNodes(netgraph, original_terms, max_nodes)
}
\arguments{
\item{netgraph}{igraph network object of ontology term IDs.}

\item{original_terms}{A character vector of ontology term IDs to ensure
coverage of.}

\item{max_nodes}{Integer to indicate number of covered nodes that makes a node
"high-traffic."}
}
\value{
A character vector of representative nodes.
}
\description{
Selects representative nodes for a network of ontology terms using igraph
clustering functionality
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoCuration")
net <- read.csv(file.path(dir, "sample_net.csv"))

netgraph <- igraph::graph_from_data_frame(d = net,
vertices = unique(unlist(net)))
original_terms <- c("NCIT:C29711", "NCIT:C270", "NCIT:C250", "NCIT:C47639",
"NCIT:C29249", "NCIT:C983", "NCIT:C247", "NCIT:C47384", "NCIT:C62002",
"NCIT:C281")

# .clusterNodes(netgraph = netgraph, original_terms = original_terms,
# max_nodes = 20)

}
\keyword{internal}
