% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodes.R
\name{.createGraph}
\alias{.createGraph}
\title{Creates igraph network object from a dataframe network representation}
\usage{
.createGraph(net, original_terms)
}
\arguments{
\item{net}{Dataframe containing a symbolic edge list of a directed network in
the first two columns. Edges are directed from the first column to the second
column. Additional columns are considered as edge attributes.}

\item{original_terms}{A character vector of ontology terms to ensure
coverage of.}
}
\value{
An igraph network object with vertex attribute "type." "type" values
include "root," "original," and "bridge."
}
\description{
Creates igraph network object from a dataframe network representation
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoCuration")
net <- read.csv(file.path(dir, "sample_net.csv"))

original_terms <- c("NCIT:C29711", "NCIT:C270", "NCIT:C250", "NCIT:C47639",
"NCIT:C29249", "NCIT:C983", "NCIT:C247", "NCIT:C47384", "NCIT:C62002",
"NCIT:C281")

# .createGraph(net = net, original_terms = original_terms)

}
\keyword{internal}
