% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnumNodes.R, R/mapNodes.R
\name{.getDescs}
\alias{.getDescs}
\title{Retrieves all original terms covered by a given node within an igraph network
object.}
\usage{
.getDescs(graph, node, original_terms)

.getDescs(graph, node, original_terms)
}
\arguments{
\item{graph}{An igraph network object.}

\item{node}{Character string; id of node to check descendants of.}

\item{original_terms}{Character vector; ids of original terms to check
coverage of.}
}
\value{
Character vector of original terms covered by "node."

Character vector of original terms covered by "node."
}
\description{
Retrieves all original terms covered by a given node within an igraph network
object.

Retrieves all original terms covered by a given node within an igraph network
object.
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoCuration")
net <- read.csv(file.path(dir, "sample_net.csv"))

graph <- igraph::graph_from_data_frame(d = net,
vertices = unique(unlist(net)))

original_terms <- c("NCIT:C29711", "NCIT:C270", "NCIT:C250", "NCIT:C47639",
"NCIT:C29249", "NCIT:C983", "NCIT:C247", "NCIT:C47384", "NCIT:C62002",
"NCIT:C281")

# .getDescs(graph = graph, node = "NCIT:C78274",
# original_terms = original_terms)

dir <- system.file("extdata", package = "OmicsMLRepoCuration")
net <- read.csv(file.path(dir, "sample_net.csv"))

graph <- igraph::graph_from_data_frame(d = net,
vertices = unique(unlist(net)))

original_terms <- c("NCIT:C29711", "NCIT:C270", "NCIT:C250", "NCIT:C47639",
"NCIT:C29249", "NCIT:C983", "NCIT:C247", "NCIT:C47384", "NCIT:C62002",
"NCIT:C281")

# .getDescs(graph = graph, node = "NCIT:C78274",
# original_terms = original_terms)

}
