% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNodes.R
\name{plotClusters}
\alias{plotClusters}
\title{Plot a network graph with multiple communities/clusters highlighted}
\usage{
plotClusters(net, communities, layout = layout_with_fr)
}
\arguments{
\item{net}{An igraph network object with a "type" vertex attribute;
"type" = "root", "bridge", "original", or "picked"}

\item{communities}{List of character vectors representing groups to mark; as
output by igraph clustering functions}

\item{layout}{An igraph layout object or layout function}
}
\value{
An igraph plot
}
\description{
Plot a network graph with multiple communities/clusters highlighted
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoCuration")
net <- read.csv(file.path(dir, "sample_net.csv"))

netgraph <- igraph::graph_from_data_frame(d = net,
vertices = unique(unlist(net)))
original_terms <- c("NCIT:C29711", "NCIT:C270", "NCIT:C250", "NCIT:C47639",
"NCIT:C29249", "NCIT:C983", "NCIT:C247", "NCIT:C47384", "NCIT:C62002",
"NCIT:C281")

V(netgraph)$type <- "bridge"
V(netgraph)[original_terms]$type <- "original"
V(netgraph)["NCIT:C1908"]$type <- "root"
V(netgraph)["NCIT:C78272"]$type <- "picked"

clusters <- list(c("NCIT:C78276", "NCIT:C471", "NCIT:C2846", "NCIT:C983",
"NCIT:C29711"), c("NCIT:C47384", "NCIT:C250", "NCIT:C276", "NCIT:C281",
"NCIT:C254"))

# plotClusters(net = netgraph, communities = clusters)

}
