% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCuratedData.R
\name{updateCuratedData}
\alias{updateCuratedData}
\title{Update a column of curated data using a new mapping schema}
\usage{
updateCuratedData(curated_data, map, column, delimiter)
}
\arguments{
\item{curated_data}{A data frame. The original version of the curated data
which needs to be updated using the new map.}

\item{map}{A data frame. The new or updated version of the map which 
corresponds to the curation of the selected dataset. The map must be 
formatted in a one-to-one schema of original to curated values.}

\item{column}{A character(1). The root name of the column which needs to be 
updated (excluding "original_" or "curated_").}

\item{delimiter}{A character(1). The delimiter used to separate values within
a single cell of the original curated data frame. Likely ";" or "<;>".}
}
\value{
A data frame of the updated curated data.
}
\description{
Update a column of curated data using a new mapping schema
}
\examples{
updateCuratedData(curated_bodysite, cBioPortal_bodysite_map, "curated_bodysite", "<;>")
updateCuratedData(curated_neonatal, cMD_feeding_map, "curated_neonatal_feeding_method", ";")

}
